# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Security Toolkit API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ScanRequest {
  /** Target URL to scan */
  url: string;
}

export type WafTestRequestCategoriesItem =
  (typeof WafTestRequestCategoriesItem)[keyof typeof WafTestRequestCategoriesItem];

export const WafTestRequestCategoriesItem = {
  sqli: "sqli",
  xss: "xss",
  cmd_injection: "cmd_injection",
  path_traversal: "path_traversal",
  xxe: "xxe",
  ssrf: "ssrf",
  lfi: "lfi",
  rfi: "rfi",
} as const;

export interface WafTestRequest {
  /** Target URL to test */
  url: string;
  /** Payload categories to test (all if omitted) */
  categories?: WafTestRequestCategoriesItem[];
}

export type WafPayloadResultStatus =
  (typeof WafPayloadResultStatus)[keyof typeof WafPayloadResultStatus];

export const WafPayloadResultStatus = {
  blocked: "blocked",
  passed: "passed",
  error: "error",
  pending: "pending",
} as const;

export interface WafPayloadResult {
  id: string;
  category: string;
  payload: string;
  status: WafPayloadResultStatus;
  statusCode?: number;
  /** Response time in milliseconds */
  responseTime?: number;
  blocked: boolean;
}

export interface WafTestResult {
  url: string;
  totalPayloads: number;
  blocked: number;
  passed: number;
  errors: number;
  blockRate: number;
  grade: string;
  results: WafPayloadResult[];
  testedAt: string;
}

export interface TechStackItem {
  name: string;
  category: string;
  version?: string;
  confidence: number;
}

export interface DnsRecord {
  type: string;
  value: string;
  ttl?: number;
}

export interface SubdomainResult {
  subdomain: string;
  ip?: string;
  status: string;
}

export type ReconResultWhois = { [key: string]: unknown };

export interface ReconResult {
  url: string;
  domain: string;
  ipAddress?: string;
  techStack: TechStackItem[];
  subdomains: SubdomainResult[];
  dnsRecords: DnsRecord[];
  whois?: ReconResultWhois;
  scannedAt: string;
}

export interface SecurityHeader {
  name: string;
  present: boolean;
  value?: string;
  score: number;
  maxScore: number;
  grade: string;
  recommendation?: string;
}

export interface HeadersAuditResult {
  url: string;
  grade: string;
  score: number;
  maxScore: number;
  headers: SecurityHeader[];
  scannedAt: string;
}

export type SpeedMetricRating =
  (typeof SpeedMetricRating)[keyof typeof SpeedMetricRating];

export const SpeedMetricRating = {
  good: "good",
  needs_improvement: "needs_improvement",
  poor: "poor",
} as const;

export interface SpeedMetric {
  name: string;
  value: number;
  unit: string;
  score: number;
  rating: SpeedMetricRating;
}

export type SpeedTestResultResourceSummary = {
  totalRequests: number;
  totalSize: number;
  htmlSize?: number;
  cssSize?: number;
  jsSize?: number;
  imageSize?: number;
};

export interface SpeedTestResult {
  url: string;
  overallScore: number;
  grade: string;
  metrics: SpeedMetric[];
  resourceSummary: SpeedTestResultResourceSummary;
  testedAt: string;
}

export type SeoIssueSeverity =
  (typeof SeoIssueSeverity)[keyof typeof SeoIssueSeverity];

export const SeoIssueSeverity = {
  error: "error",
  warning: "warning",
  info: "info",
} as const;

export interface SeoIssue {
  type: string;
  severity: SeoIssueSeverity;
  message: string;
  element?: string;
}

export type SeoAuditResultOpenGraph = { [key: string]: unknown };

export type SeoAuditResultTwitterCard = { [key: string]: unknown };

export type SeoAuditResultStructuredDataItem = { [key: string]: unknown };

export type SeoAuditResultHeadings = { [key: string]: unknown };

export interface SeoAuditResult {
  url: string;
  score: number;
  grade: string;
  title?: string;
  description?: string;
  canonicalUrl?: string;
  robotsMeta?: string;
  openGraph?: SeoAuditResultOpenGraph;
  twitterCard?: SeoAuditResultTwitterCard;
  structuredData?: SeoAuditResultStructuredDataItem[];
  headings?: SeoAuditResultHeadings;
  issues: SeoIssue[];
  internalLinks?: number;
  externalLinks?: number;
  brokenLinks?: number;
  imagesWithoutAlt?: number;
  scannedAt: string;
}

export type ExportRequestFormat =
  (typeof ExportRequestFormat)[keyof typeof ExportRequestFormat];

export const ExportRequestFormat = {
  html: "html",
  json: "json",
} as const;

export type ExportRequestData = { [key: string]: unknown };

export interface ExportRequest {
  format: ExportRequestFormat;
  data: ExportRequestData;
  title?: string;
}

export interface ExportResult {
  format: string;
  content: string;
  filename: string;
}