# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Security Toolkit API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recon/scan:
    post:
      operationId: runReconScan
      tags: [recon]
      summary: Run reconnaissance scan
      description: Identify tech stack, subdomains, and DNS records
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
      responses:
        "200":
          description: Reconnaissance results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconResult"
  /headers/audit:
    post:
      operationId: runHeadersAudit
      tags: [headers]
      summary: Audit security headers
      description: Check and grade HTTP security headers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
      responses:
        "200":
          description: Header audit results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeadersAuditResult"
...
components:
  schemas:
...
    ScanRequest:
      type: object
      properties:
        url:
          type: string
          description: Target URL to scan
      required:
        - url
...
    TechStackItem:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        version:
          type: string
        confidence:
          type: integer
      required:
        - name
        - category
        - confidence
    DnsRecord:
      type: object
      properties:
        type:
          type: string
        value:
          type: string
        ttl:
          type: integer
      required:
        - type
        - value
    SubdomainResult:
      type: object
      properties:
        subdomain:
          type: string
        ip:
          type: string
        status:
          type: string
      required:
        - subdomain
        - status
    ReconResult:
      type: object
      properties:
        url:
          type: string
        domain:
          type: string
        ipAddress:
          type: string
        techStack:
          type: array
          items:
            $ref: "#/components/schemas/TechStackItem"
        subdomains:
          type: array
          items:
            $ref: "#/components/schemas/SubdomainResult"
        dnsRecords:
          type: array
          items:
            $ref: "#/components/schemas/DnsRecord"
        whois:
          type: object
          additionalProperties: true
        scannedAt:
          type: string
      required:
        - url
        - domain
        - techStack
        - subdomains
        - dnsRecords
        - scannedAt
    SecurityHeader:
      type: object
      properties:
        name:
          type: string
        present:
          type: boolean
        value:
          type: string
        score:
          type: integer
        maxScore:
          type: integer
        grade:
          type: string
        recommendation:
          type: string
      required:
        - name
        - present
        - score
        - maxScore
        - grade
    HeadersAuditResult:
      type: object
      properties:
        url:
          type: string
        grade:
          type: string
        score:
          type: integer
        maxScore:
          type: integer
        headers:
          type: array
          items:
            $ref: "#/components/schemas/SecurityHeader"
        scannedAt:
          type: string
      required:
        - url
        - grade
        - score
        - maxScore
        - headers
        - scannedAt
...