# lib/api-zod/src/generated/types/securityHeader.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Security Toolkit API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SecurityHeader {
  name: string;
  present: boolean;
  value?: string;
  score: number;
  maxScore: number;
  grade: string;
  recommendation?: string;
}