# lib/api-zod/src/generated/types/subdomainResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Security Toolkit API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubdomainResult {
  subdomain: string;
  ip?: string;
  status: string;
}