# lib/api-zod/src/generated/types/techStackItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Security Toolkit API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TechStackItem {
  name: string;
  category: string;
  version?: string;
  confidence: number;
}